+++ src/lib.rs
@@ -2,16 +2,16 @@
 
 pub use crate::array::H3Converter;
 pub use crate::axis::AxisOrder;
+pub use crate::coverage::CellCoverage;
 pub use crate::error::Error;
 pub use crate::resolution::ResolutionSearchMode;
-pub use crate::set::CellSet;
 
 pub use crate::transform::Transform;
 
 mod array;
 mod axis;
+mod coverage;
 mod error;
 mod resolution;
-mod set;
 mod sphere;
 mod transform;